recognition/tasks/partialface/requirements.txt
dareblopy
torchjpeg